fixed some warnings
git-svn-id: http://vtp.googlecode.com/svn/trunk@220 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainSDK/vtlib/core/LocalProjection.cpp
@@ -31,7 +31,7 @@ void vtLocalProjection::SetDegreeOrigin(const DPoint2 &degrees)
 	m_EarthOrigin = degrees;
 
 	// estimate meters per degree of longitude, using the terrain origin
-	m_fMetersPerLongitude = EstimateDegreesToMeters(degrees.y);
+	m_fMetersPerLongitude = (float) EstimateDegreesToMeters(degrees.y);
 }
 
 void vtLocalProjection::SetMeterOrigin(const DPoint2 &meters)
